requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aenum==3.1.15
aioboto3==13.3.0
aiobotocore==2.16.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
anytree==2.12.1
APScheduler==3.11.0
argcomplete==1.10.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.3.0
audioread==3.0.1
autograd==1.7.0
babel==2.16.0
beartype==0.18.5
beautifulsoup4==4.8.2
bidict==0.23.1
bleach==6.2.0
blis==0.7.11
boto3==1.35.81
botocore==1.35.81
bs4==0.0.2
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cmudict==1.0.32
comm==0.2.2
compressed_rtf==1.0.6
confection==0.1.5
contourpy==1.3.1
coqpit-config==0.1.2
coqui-tts==0.25.1
coqui-tts-trainer==0.2.2
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
dataclasses-json==0.6.7
datasets==3.2.0
dateparser==1.1.8
debugpy==1.8.11
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
Distance==0.1.3
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
einops==0.8.0
emoji==1.2.0
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.1.0
extract-msg==0.28.7
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gensim==4.3.3
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.1.1
gremlinpython==3.7.3
grpcio==1.67.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-jamo==1.0.1
hangul-romanize==0.1.0
hnswlib==0.8.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
hyppo==0.4.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
IMAPClient==2.1.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
jamo==0.4.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==1.2.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kollocate==0.0.2
koparadigm==0.10.0
koreanize-matplotlib==0.1.1
kss==6.0.4
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
langcodes==3.5.0
langsmith==0.2.10
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.2.post1
lightrag-hku==1.1.1
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.11
mistune==3.1.0
monotonic-alignment-search==0.1.1
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
neo4j==5.27.0
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
olefile==0.47
ollama==0.4.6
openai==1.59.7
oracledb==2.5.1
orjson==3.10.14
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pdfminer.six==20191110
pecab==1.0.8
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
POT==0.9.5
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pychrome==0.2.4
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pymilvus==2.5.3
pymongo==4.10.1
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.1
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python-pptx==0.6.23
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
selenium==4.27.1
Send2Trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.12.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.37
srsly==2.5.0
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
SudachiDict-core==20241021
SudachiPy==0.6.10
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
textract==1.6.5
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.2.1
torch==2.5.0+cu118
torchaudio==2.5.0+cu118
torchvision==0.20.0+cu118
tornado==6.4.2
tossi==0.3.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.2
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
trl==0.13.0
typeguard==4.4.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
umap-learn==0.5.7
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
Whoosh==2.7.4
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
